Rename, remove whitespace, remove variables
