+++ db/seeds/dev/movie_seed.js
@@ -1,6 +1,6 @@
-const movieData = require('../../data/movies_data.js');
+const movieData = require("../../data/movies_data.js");
 const createMovie = async (knex, movie) => {
-  await knex('movies').insert({
+  await knex("movies").insert({
     title: movie.title.title,
     genre: movie.genres,
     content_rating: movie.certificates && movie.certificates.US[0].certificate,
@@ -11,16 +11,16 @@ const createMovie = async (knex, movie) => {
     release_date: movie.releaseDate,
     runtime: movie.title.runningTimeInMinutes,
   });
-}
+};
 exports.seed = async (knex) => {
   try {
-    await knex('movies').del()
-    let allMovieData = await movieData(category)
-    let moviePromises = allMovieData.map(movie => {
+    await knex("movies").del();
+    let allMovieData = await movieData(category);
+    let moviePromises = allMovieData.map((movie) => {
       return createMovie(knex, movie);
     });
     return Promise.all(moviePromises);
   } catch (error) {
-    console.log(`Error seeding data: ${error}`)
+    console.log(`Error seeding data: ${error}`);
   }
 };
